Реализация таблицы приложения

+++ PhoneDirectoryParser/PhoneDirectory.cs
@@ -15,7 +15,7 @@ namespace PhoneDirectoryParser
             CalledNumber = long.Parse(calledNumber);
             DateAndTimeToStartAConversation = DateTime.Parse(dateAndTimeToStartAConversation);
             DurationOfMinutes = int.Parse(durationOfMinutes);
-            Cost = double.Parse(cost);
+            Cost = decimal.Parse(cost);
         }
         public PhoneDirectory(params string[] callLine) {
             PhoneNumber = long.Parse(callLine[0]);
@@ -24,7 +24,7 @@ namespace PhoneDirectoryParser
             CalledNumber = long.Parse(callLine[3]);
             DateAndTimeToStartAConversation = DateTime.Parse(callLine[4]);
             DurationOfMinutes = int.Parse(callLine[5]);
-            Cost = double.Parse(callLine[6], System.Globalization.NumberStyles.AllowDecimalPoint, System.Globalization.NumberFormatInfo.InvariantInfo);
+            Cost = decimal.Parse(callLine[6], System.Globalization.NumberStyles.AllowDecimalPoint, System.Globalization.NumberFormatInfo.InvariantInfo);
         }
 
         public long PhoneNumber { get; set; }
@@ -39,6 +39,6 @@ namespace PhoneDirectoryParser
 
         public int DurationOfMinutes { get; set; }
 
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
     }
 }
